v1.2.0: bug修复 + DMG拖拽安装支持
修复:
- BUG-001: 图表强制解包 as! -> guard let as? (ChargeChartView, FanView)
- BUG-002: 删除 FanView.setSliderValue 死代码
- BUG-003: MetricBarView.Entry 加 showStatus 标志，消除 warnAt:2 hack
- BUG-004: smc.c 2字节有符号数用 int16_t 正确扩展
- BUG-005: 删除 fanhelper.c print_backlight 死代码
- BUG-006: DESIGN.md barY 公式修正

改进:
- build.sh 版本号变量化 + 集成 DMG 打包 (Finder alias + 窗口布局)
- 新增 BUGFIXES.md / CHANGELOG.md

+++ smc.c
@@ -100,7 +100,7 @@ int smc_read_int(const char *key, int *out) {
         *out = ((unsigned char)b[0] << 24) | ((unsigned char)b[1] << 16) |
                ((unsigned char)b[2] << 8)  |  (unsigned char)b[3];
     } else if (ki.dataSize == 2) {
-        *out = ((unsigned char)b[0] << 8) | (unsigned char)b[1];
+        *out = (int16_t)(((unsigned char)b[0] << 8) | (unsigned char)b[1]);
     } else {
         *out = (unsigned char)b[0];
     }
